Removed Materials.
Models now hold on to the Texture and Shader themselves. The benefit of Materials is that they allow for easy reuse of the same shader + texture pair but the drawback is that each pair of shader + texture must be explicitly defined.

When creating materials with multiple textures, I may consider adding a SetMaterial function that sets each texture of a model accordingly.