Completed Multi-Pass Lighting + Ambient
Scene can have multiple lights now and will have an ambient light color that is used during the first pass over the scene.

One potential issue is that the ambient pass samples the texture once and the light pass will sample it again which is not ideal. I'm not sure how to fix this yet.