Added support for other light types.
The LightStruct is now named LightData and has data for directional, point, and spot lights.

However, defaultlightshader still assumes all lights are directional. With the type field of LightData, we can conditionally render different types of lights using the data stored in LightData.

LightData ends up being 64 bytes to account for all of the possible information. Which is also the exact size of a vertex.