Began work on loading texture meta data from file
Textures are soon to hold a SamplerState instead of shaders holding them. As such, the meta data for textures containing just the sampler desc data (for now) will be loaded from a .texturemeta file. Eventually, shaders, sprites, many other things will get this treatment as well but for now it is just for textures.