Reworked Samplers and Render2DSystem
Samplers are limited heavily in the shaders so instead of each texture holding a unique sampler, a sampler is only generated per unique samplerDesc, and all samplers are held in a cache managed by the textureManager. Each texture still has a pointer to a sampler but the pointers will hold the same values across textures with the same sampler settings.

As such, batching is limited as well by the sampler type so this has been adjusted and each Vertex2D also holds a samplerIndex now too.