Scaled Back Sprite Batcher
Due to an oversight, I did not realize that dynamically addressing the indices of an array of textures or samplers was impossible in DirectX11 unless you are using a Texture2DArray, for now I have decided to just scale back drastically such that each batch may only have 1 texture/sampler combo. Which is far from ideal but batching may still be done through sprite atlases.

In the future I will revisit this problem to find a more graceful solution.