Added SpriteAnimation and way to load them
SpriteAnimations are just a vector of SpriteAnimationFrames, the SpriteAnimator component will take a SpriteAnimation and the SpriteAnimationSystem will update Sprites based off of their animations.

Sprite animations are loaded from a .spriteanimation file where each frame or block of frames is denoted by "#FRAME".

Under #FRAME you can specify these fields:

Count: how many frames are in the chunk.
Slide: the source origin offset of each successive frame.
Texture: the name of the texture this frame chunk uses.
Dimensions: the dimensions of the sprite from the source.
Source_Origin: the origin on the source image of the first sprite.
Duration: how long the chunk will be displayed, for n frames, each frame will be displayed for duration/n seconds.