Refactored Render2DSystem
There is now a Renderer2D which you can submit a sprite to and it directly takes the sprites fields as input.

This is so that the Text system can use the Renderer2D without directly having multiple Sprite components.

In the future, I'd like to create a Renderer3D and then have an abstraction layer for a general Renderer which can swap between a deferred renderer and forward renderer as well as Renderers for other graphics APIs if I get that far.