Switched sprite submission to use a builder.
Instead of tons of variations on SubmitSprite, the Renderer2D now has a BuildSprite function which will return a SpriteData. The caller can then build the SpriteData they received and when done call Submit on it to submit the sprite to the Renderer.

This is not the most elegant solution, if I think of a better way to avoid this issue I will come back to this.