Introduced SpriteData
Went back on builder pattern and decided to use a SpriteData struct which holds the shader, texture, sourceRect, and tint of a sprite.

The Renderer2D takes a const reference to a SpriteData and some data for where to draw the sprite. There are only 4 definitions for SubmitSprite now and less copying.