Refactored Resources further.
Resources like Mesh's (VBO's and IBO's) and Texture2D's are now not uploaded to the GPU on construction. They instead store their data CPU side and must be manually uploaded to the GPU before use using Upload() passing the device (and devicecontext for Texture2D). Likewise, the GPU data can be freed by calling Release().

This is to preserve video memory by not sending unnecessary data to the GPU. For example: The skull mesh is not currently being used so there is no reason to send it to the GPU. Storing data CPU side also lets me create binary files for this data much more easily.

As well, I am further reworking Resources to use an all purpose ResourceManager that will handle all resources. The motivation being that I do not want the user to see the specific functions for loading different file types or to manually determine when to use the binary file.